Move BuildingTracerModule to its own package
